refactor(pipeline): extract build_supergraph_only for the test harness
test_harness::build_common hand-replayed build_pipeline's sequence to
obtain a supergraph-only composition, so any change to the acquire or
assemble steps had to be mirrored there by hand. The sequence now
lives in build_supergraph_only beside build_pipeline. The two do not
share code: build_pipeline's acquire also runs telemetry, APQ Redis,
and the persisted-query manifest before the activation point of no
return, and a shared helper activating plugins mid-sequence would
force fallible work past activation. Adjacency plus a doc note is the
drift protection.
